Global refactoring
 * Extracting ui (views and components)
 * Extracting core functions
 * redesigning dependencies (especially gi)
