+++ gtimelog
@@ -1,16 +1,14 @@
-#!/usr/bin/python3
+#!/usr/bin/env python
 """
 Script to run GTimeLog from the source checkout without installing
 """
 import os
 import sys
 
+from gtimelog.main import main
 
 basedir = os.path.dirname(os.path.realpath(__file__))
 pkgdir = os.path.join(basedir, 'src')
 sys.path.insert(0, pkgdir)
-
-from gtimelog.main import main  # noqa: E402
-
 
 main()
