+++ src/gtimelog/debian-paths.py
@@ -2,12 +2,12 @@
 Resource locations for running out of Debian package installs
 """
 
-UI_FILE = '/usr/share/gtimelog/gtimelog.ui'
-PREFERENCES_UI_FILE = '/usr/share/gtimelog/preferences.ui'
-ABOUT_DIALOG_UI_FILE = '/usr/share/gtimelog/about.ui'
-SHORTCUTS_UI_FILE = '/usr/share/gtimelog/shortcuts.ui'
-MENUS_UI_FILE = '/usr/share/gtimelog/menus.ui'
-CSS_FILE = '/usr/share/gtimelog/gtimelog.css'
+UI_FILE = '/usr/share/gtimelog/ui/views/gtimelog.ui'
+PREFERENCES_UI_FILE = '/usr/share/gtimelog/ui/views/preferences.ui'
+ABOUT_DIALOG_UI_FILE = '/usr/share/gtimelog/ui/views/about.ui'
+SHORTCUTS_UI_FILE = '/usr/share/gtimelog/ui/views/shortcuts.ui'
+MENUS_UI_FILE = '/usr/share/gtimelog/ui/views/menus.ui'
+CSS_FILE = '/usr/share/gtimelog/ui/views/gtimelog.css'
 
 LOCALE_DIR = '/usr/share/locale'
 CONTRIBUTORS_FILE = '/usr/share/doc/gtimelog/CONTRIBUTORS.rst'
