+++ src/gtimelog/tests/__main__.py
@@ -1,5 +1,4 @@
 from gtimelog.tests import main
 
-
 if __name__ == '__main__':
     main()
