+++ src/gtimelog/tests/core/test_timelog.py
@@ -1,0 +1,319 @@
+import doctest
+from datetime import datetime, time
+import unittest
+import sys
+from io import StringIO
+
+import freezegun
+from unittest import mock
+
+from gtimelog.core.exports import Exports
+from gtimelog.tests.commons import Checker
+from gtimelog.tests.core import make_time_window
+
+
+def doctest_timewindow_repr():
+    """Test for TimeWindow.__repr__
+
+        >>> minimum = datetime(2013, 12, 3)
+        >>> maximum = datetime(2013, 12, 4)
+        >>> vm = time(2, 0)
+
+        >>> make_time_window(minimum=minimum, maximum=maximum, vm=vm)
+        <TimeWindow: 2013-12-03 00:00:00..2013-12-04 00:00:00>
+
+    """
+
+
+def doctest_timewindow_reread_no_file():
+    """Test for TimeWindow.reread
+
+        >>> window = make_time_window('/nosuchfile')
+
+    There's no error.
+
+        >>> len(window.items)
+        0
+        >>> window.last_time()
+
+    """
+
+
+def doctest_timewindow_reread_bad_timestamp():
+    """Test for TimeWindow.reread
+
+        >>> minimum = datetime(2013, 12, 4)
+        >>> maximum = datetime(2013, 12, 5)
+        >>> vm = time(2, 0)
+
+        >>> sampledata = StringIO('''
+        ... 2013-12-04 09:00: start **
+        ... # hey: this is not a timestamp
+        ... 2013-12-04 09:14: gtimelog: write some tests
+        ... ''')
+
+        >>> window = make_time_window(sampledata, minimum, maximum, vm)
+
+    There's no error, the line with a bad timestamp is silently skipped.
+
+        >>> len(window.items)
+        2
+
+    """
+
+
+def doctest_timewindow_reread_bad_ordering():
+    """Test for TimeWindow.reread
+
+        >>> minimum = datetime(2013, 12, 4)
+        >>> maximum = datetime(2013, 12, 5)
+
+        >>> sampledata = StringIO('''
+        ... 2013-12-04 09:00: start **
+        ... 2013-12-04 09:14: gtimelog: write some tests
+        ... 2013-12-04 09:10: gtimelog: whoops clock got all confused
+        ... 2013-12-04 09:10: gtimelog: so this will need to be fixed
+        ... ''')
+
+        >>> window = make_time_window(sampledata, minimum, maximum)
+
+    There's no error, the timestamps have been reordered, but note that
+    order was preserved for events with the same timestamp
+
+        >>> for t, e in window.items:
+        ...     print("%s: %s" % (t.strftime('%H:%M'), e))
+        09:00: start **
+        09:10: gtimelog: whoops clock got all confused
+        09:10: gtimelog: so this will need to be fixed
+        09:14: gtimelog: write some tests
+
+        >>> window.last_time()
+        datetime.datetime(2013, 12, 4, 9, 14)
+
+    """
+
+
+def doctest_timewindow_count_days():
+    """Test for TimeWindow.count_days
+
+        >>> minimum = datetime(2013, 12, 2)
+        >>> maximum = datetime(2013, 12, 9)
+        >>> vm = time(2, 0)
+
+        >>> sampledata = StringIO('''
+        ... 2013-12-04 09:00: start **
+        ... 2013-12-04 09:14: gtimelog: write some tests
+        ... 2013-12-04 09:10: gtimelog: whoops clock got all confused
+        ... 2013-12-04 09:10: gtimelog: so this will need to be fixed
+        ...
+        ... 2013-12-05 22:30: some fictional late night work **
+        ... 2013-12-06 00:30: frobnicate the widgets
+        ...
+        ... 2013-12-08 09:00: work **
+        ... 2013-12-08 09:01: and stuff
+        ... ''')
+
+        >>> window = make_time_window(sampledata, minimum, maximum, vm)
+        >>> window.count_days()
+        3
+
+    """
+
+
+def doctest_timewindow_last_entry():
+    """Test for TimeWindow.last_entry
+
+        >>> window = make_time_window()
+
+    Case #1: no items
+
+        >>> window.items = []
+        >>> window.last_entry()
+
+    Case #2: single item
+
+        >>> window.items = [
+        ...     (datetime(2013, 12, 4, 9, 0), 'started **'),
+        ... ]
+        >>> start, stop, duration, tags, entry = window.last_entry()
+        >>> start == stop == datetime(2013, 12, 4, 9, 0)
+        True
+        >>> duration
+        datetime.timedelta(0)
+        >>> entry
+        'started **'
+
+    Case #3: single item at start of new day
+
+        >>> window.items = [
+        ...     (datetime(2013, 12, 3, 12, 0), 'stuff'),
+        ...     (datetime(2013, 12, 4, 9, 0), 'started **'),
+        ... ]
+        >>> start, stop, duration, tags, entry = window.last_entry()
+        >>> start == stop == datetime(2013, 12, 4, 9, 0)
+        True
+        >>> duration
+        datetime.timedelta(0)
+        >>> entry
+        'started **'
+
+
+    Case #4: several items
+
+        >>> window.items = [
+        ...     (datetime(2013, 12, 4, 9, 0), 'started **'),
+        ...     (datetime(2013, 12, 4, 9, 31), 'gtimelog: tests'),
+        ... ]
+        >>> start, stop, duration, tags, entry = window.last_entry()
+        >>> start
+        datetime.datetime(2013, 12, 4, 9, 0)
+        >>> stop
+        datetime.datetime(2013, 12, 4, 9, 31)
+        >>> duration
+        datetime.timedelta(0, 1860)
+        >>> entry
+        'gtimelog: tests'
+
+    """
+
+
+def doctest_Exports_to_csv_complete():
+    r"""Tests for Exports.to_csv_complete
+
+        >>> minimum = datetime(2008, 6, 1)
+        >>> maximum = datetime(2008, 7, 1)
+        >>> vm = time(2, 0)
+
+        >>> sampledata = StringIO('''
+        ... 2008-06-03 12:45: start
+        ... 2008-06-03 13:00: something
+        ... 2008-06-03 14:45: something else
+        ... 2008-06-03 15:45: etc
+        ... 2008-06-05 12:45: start
+        ... 2008-06-05 13:15: something
+        ... 2008-06-05 14:15: rest **
+        ... 2008-06-05 16:15: let's not mention this ever again ***
+        ... ''')
+
+        >>> window = make_time_window(sampledata, minimum, maximum, vm)
+
+        >>> Exports(window).to_csv_complete(sys.stdout)
+        task,time (minutes)
+        etc,60
+        something,45
+        something else,105
+
+    """
+
+
+def doctest_Exports_to_csv_daily():
+    r"""Tests for Exports.to_csv_daily
+
+        >>> minimum = datetime(2008, 6, 1)
+        >>> maximum = datetime(2008, 7, 1)
+        >>> vm = time(2, 0)
+
+        >>> sampledata = StringIO('''
+        ... 2008-06-03 12:45: start
+        ... 2008-06-03 13:00: something
+        ... 2008-06-03 14:45: something else
+        ... 2008-06-03 15:45: etc
+        ... 2008-06-05 12:45: start
+        ... 2008-06-05 13:15: something
+        ... 2008-06-05 14:15: rest **
+        ... ''')
+
+        >>> window = make_time_window(sampledata, minimum, maximum, vm)
+
+        >>> Exports(window).to_csv_daily(sys.stdout)
+        date,day-start (hours),slacking (hours),work (hours)
+        2008-06-03,12.75,0.0,3.0
+        2008-06-04,0.0,0.0,0.0
+        2008-06-05,12.75,1.0,0.5
+
+    """
+
+
+def doctest_Exports_icalendar():
+    r"""Tests for Exports.icalendar
+
+        >>> minimum = datetime(2008, 6, 1)
+        >>> maximum = datetime(2008, 7, 1)
+        >>> vm = time(2, 0)
+
+        >>> sampledata = StringIO(r'''
+        ... 2008-06-03 12:45: start **
+        ... 2008-06-03 13:00: something
+        ... 2008-06-03 15:45: something, else; with special\chars
+        ... 2008-06-05 12:45: start **
+        ... 2008-06-05 13:15: something
+        ... 2008-06-05 14:15: rest **
+        ... ''')
+
+        >>> window = make_time_window(sampledata, minimum, maximum, vm)
+
+        >>> with freezegun.freeze_time("2015-05-18 15:40"):
+        ...     with mock.patch('socket.getfqdn') as mock_getfqdn:
+        ...         mock_getfqdn.return_value = 'localhost'
+        ...         Exports(window).icalendar(sys.stdout)
+        ... # doctest: +REPORT_NDIFF
+        BEGIN:VCALENDAR
+        PRODID:-//gtimelog.org/NONSGML GTimeLog//EN
+        VERSION:2.0
+        BEGIN:VEVENT
+        UID:be5f9be205c2308f7f1a30d6c399d6bd@localhost
+        SUMMARY:start **
+        DTSTART:20080603T124500
+        DTEND:20080603T124500
+        DTSTAMP:20150518T154000Z
+        END:VEVENT
+        BEGIN:VEVENT
+        UID:33c7e212fed11eda71d5acd4bd22119b@localhost
+        SUMMARY:something
+        DTSTART:20080603T124500
+        DTEND:20080603T130000
+        DTSTAMP:20150518T154000Z
+        END:VEVENT
+        BEGIN:VEVENT
+        UID:b10c11beaf91df16964a46b4c87420b1@localhost
+        SUMMARY:something\, else\; with special\\chars
+        DTSTART:20080603T130000
+        DTEND:20080603T154500
+        DTSTAMP:20150518T154000Z
+        END:VEVENT
+        BEGIN:VEVENT
+        UID:04964eef67ec22178d74fe4c0f06aa2a@localhost
+        SUMMARY:start **
+        DTSTART:20080605T124500
+        DTEND:20080605T124500
+        DTSTAMP:20150518T154000Z
+        END:VEVENT
+        BEGIN:VEVENT
+        UID:2b51ea6d1c26f02d58051a691657068d@localhost
+        SUMMARY:something
+        DTSTART:20080605T124500
+        DTEND:20080605T131500
+        DTSTAMP:20150518T154000Z
+        END:VEVENT
+        BEGIN:VEVENT
+        UID:bd6bfd401333dbbf34fec941567d5d06@localhost
+        SUMMARY:rest **
+        DTSTART:20080605T131500
+        DTEND:20080605T141500
+        DTSTAMP:20150518T154000Z
+        END:VEVENT
+        END:VCALENDAR
+
+    """
+
+
+def additional_tests(): # for setup.py
+    return doctest.DocTestSuite(optionflags=doctest.NORMALIZE_WHITESPACE,
+                                checker=Checker())
+
+
+def test_suite():
+    return unittest.TestSuite([
+        unittest.defaultTestLoader.loadTestsFromName(__name__),
+        additional_tests(),
+    ])
