+++ src/gtimelog/ui/views/about.ui
