+++ src/gtimelog/ui/views/gtimelog-large.png
