+++ src/gtimelog/ui/views/gtimelog.css
