+++ src/gtimelog/ui/views/gtimelog.png
