+++ src/gtimelog/ui/views/gtimelog.ui
