+++ src/gtimelog/ui/views/menus.ui
