+++ src/gtimelog/ui/views/preferences.ui
