+++ src/gtimelog/ui/views/shortcuts.ui
