+++ tox.ini
@@ -7,6 +7,7 @@ setenv =
     LC_ALL=C
 deps =
     freezegun
+;    PyGObject
 skip_install = true
 commands =
     pip install --no-deps -e .
@@ -16,6 +17,7 @@ commands =
 basepython = python3
 deps =
     freezegun
+;    PyGObject
     coverage
     -cconstraints.txt
 skip_install = true
@@ -27,6 +29,7 @@ commands =
 [testenv:py]
 deps =
     freezegun
+;    PyGObject
 skip_install = true
 commands =
     python --version
